TownHall Production & Experience (pt.1)

+++ Assets/Project/Scripts/EventBus/Events/ResourceBalanceChangedEvent.cs
@@ -3,8 +3,8 @@ namespace Events
     public class ResourceBalanceChangedEvent
     {
         public readonly ResourceType ResourceType;
-        public readonly int Amount;
-        public ResourceBalanceChangedEvent(ResourceType type, int amount)
+        public readonly float Amount;
+        public ResourceBalanceChangedEvent(ResourceType type, float amount)
         {
             ResourceType = type;
             Amount = amount;
