+++ Assets/Project/Scripts/EventBus/Events/ResourceSpendEvent.cs
@@ -3,8 +3,8 @@ namespace Events
     public class ResourceSpendEvent
     {
         public readonly ResourceType ResourceType;
-        public readonly int Amount;
-        public ResourceSpendEvent(ResourceType type, int amount)
+        public readonly float Amount;
+        public ResourceSpendEvent(ResourceType type, float amount)
         {
             ResourceType = type;
             Amount = amount;
